+++ src/config.ts
@@ -1,3 +1,4 @@
 export const config = {
   COLLECTOR_VERSION: '2.2.0',
+  ERRORS_TERMINATE_TRACKING: [403, 409, 422],
 }